updated with java streams

+++ backendservice/src/main/java/com/google/gpay/virtualqueue/backendservice/proto/GetShopResponse.java
@@ -13,6 +13,8 @@ limitations under the License.
 
 package com.google.gpay.virtualqueue.backendservice.proto;
 
+import java.util.concurrent.atomic.AtomicInteger;
+
 import com.google.gpay.virtualqueue.backendservice.model.Shop;
 
 import lombok.AllArgsConstructor;
@@ -22,5 +24,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class GetShopResponse {
     private Shop shop;
-    private long customersInQueue;
+    private AtomicInteger customersInQueue;
 }
